pyproject.toml
[tool.poetry]
name = "aggr-bot"
version = "0.1.0"
description = "Test bot to aggregate data from mondodb on user query"
authors = ["chrnv chernov.sib@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^4.5.0"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
